pyproject.toml
[project]
name = "interview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "accelerate>=1.8.1",
    "django>=5.2.4",
    "django-environ==0.12.0",
    "djangorestframework>=3.16.0",
    "dnspython==2.3.0",
    "ffmpeg>=1.4",
    "gensim>=4.3.3",
    "grandalf>=0.8",
    "huggingface-hub>=0.33.2",
    "ipython>=9.4.0",
    "langchain-huggingface>=0.3.0",
    "langgraph>=0.5.1",
    "nltk>=3.9.1",
    "openai-whisper>=20250625",
    "pgadmin4>=9.3",
    "psycopg2>=2.9.10",
    "pymongo==4.6.3",
    "python-dotenv>=1.1.1",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.53.1",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu128"

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu123"
